+++ packages/webpack-cli/lib/index.js
@@ -1,5 +1,3 @@
 const CLI = require("./webpack-cli");
-const utils = require("./utils");
 
 module.exports = CLI;
-module.exports.utils = utils;
