--- packages/webpack-cli/lib/utils/capitalize-first-letter.js
@@ -1,9 +1,0 @@
-const capitalizeFirstLetter = (string) => {
-  if (typeof string !== "string") {
-    return "";
-  }
-
-  return string.charAt(0).toUpperCase() + string.slice(1);
-};
-
-module.exports = capitalizeFirstLetter;
