--- packages/webpack-cli/lib/utils/run-command.js
@@ -1,13 +1,0 @@
-const execa = require("execa");
-const utils = require("./index");
-
-async function runCommand(command, args = []) {
-  try {
-    await execa(command, args, { stdio: "inherit", shell: true });
-  } catch (error) {
-    utils.logger.error(error.message);
-    process.exit(2);
-  }
-}
-
-module.exports = runCommand;
