--- packages/webpack-cli/lib/utils/to-kebab-case.js
@@ -1,5 +1,0 @@
-const toKebabCase = (str) => {
-  return str.replace(/([a-z0-9])([A-Z])/g, "$1-$2").toLowerCase();
-};
-
-module.exports = toKebabCase;
